+++ SOFTWARE ENGINEERING/eRegistrar/src/main/java/edu/miu/cs/cs425/eregistrar/eregistrar/repository/ClassroomRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 
 
-public interface ClassroomRepository extends CrudRepository<Classroom, Long> {
+public interface ClassroomRepository extends CrudRepository<Classroom, Integer> {
 
 }
